Remove classes not needed anymore
